Adds null checking to data.result
In certain situations, data.result can equal "null" instead of "undefined" and this check covers both cases.  In our case, we had a bad field name in a query, and it was successfully parsing the fault code, but then getting a TypeError on this line because data.result was "null" not "undefined".